Make Hastebin uploading configurable
