+++ src/main/java/net/modfest/utilities/ModFestUtilities.java
@@ -2,7 +2,6 @@ package net.modfest.utilities;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.google.gson.annotations.Expose;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.requests.GatewayIntent;
@@ -17,10 +16,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import javax.security.auth.login.LoginException;
-import java.net.URI;
 import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
 import java.util.List;
 
 public class ModFestUtilities implements ModInitializer {
@@ -81,27 +77,5 @@ public class ModFestUtilities implements ModInitializer {
             discord.shutdown();
             discord = null; // allow garbage collection, as the event listener has a reference to the MinecraftServer.
         }
-    }
-
-    public static void handleCrashReport(String report) {
-        LOGGER.info("[ModFest] Publishing crash report.");
-        try {
-            HttpResponse<String> response = CLIENT.send(HttpRequest.newBuilder()
-                    .uri(URI.create("https://hastebin.com/documents"))
-                    .POST(HttpRequest.BodyPublishers.ofString(report))
-                    .header("Content-Type", "text/html; charset=UTF-8")
-                    .build(), HttpResponse.BodyHandlers.ofString());
-            if(response.statusCode() / 100 != 2) throw new RuntimeException("Non-success status code from Hastebin request " + response);
-            
-            HasteBinResponse haste = GSON.fromJson(response.body(), HasteBinResponse.class);
-            LOGGER.info("[ModFest] Crash report available at: https://hastebin.com/" + haste.key);
-            WebHookJson.createSystem("The server has crashed!\nReport: https://hastebin.com/" + haste.key).send().join();
-        } catch (Exception e) {
-            ModFestUtilities.LOGGER.error("[ModFest] Crash log failed to send.", e);
-        }
-    }
-
-    public static class HasteBinResponse {
-        @Expose public String key = "";
     }
 }
