+++ src/main/java/net/modfest/utilities/mixin/CrashReportMixin.java
@@ -1,7 +1,7 @@
 package net.modfest.utilities.mixin;
 
 import net.minecraft.util.crash.CrashReport;
-import net.modfest.utilities.ModFestUtilities;
+import net.modfest.utilities.CrashUploader;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
@@ -12,11 +12,10 @@ import java.io.File;
 
 @Mixin(CrashReport.class)
 public abstract class CrashReportMixin {
-
     @Shadow public abstract String asString();
 
     @Inject(method = "writeToFile", at = @At("RETURN"))
     private void writeToFile(File file, CallbackInfoReturnable<Boolean> info) {
-        ModFestUtilities.handleCrashReport(this.asString());
+        CrashUploader.handleCrashReport(this.asString());
     }
 }
